fix(core): P3 安全批次 Phase 1 — 溢出与线程安全修复
- class_group factor_ideal/factor_principal_ideal: __int128_t 防止 b*r 溢出
- class_group val=0 边界: a=b*r 时正确计算 v_P = v_p(b) + 1
- SieveParams::combined_threshold() 返回 uint16_t 防止加法溢出
- SparseRow::test() 移除 const_cast，未排序时用线程安全的线性扫描

+++ include/gnfs/sieve/lattice_sieve.hpp
@@ -35,9 +35,9 @@ struct SieveParams {
     bool enable_2lp = true;                    // 启用 2LP (two large primes)
     bool enable_3lp = false;                   // 启用 3LP (three large primes)
 
-    /// 计算合并阈值
-    [[nodiscard]] uint8_t combined_threshold() const noexcept {
-        return rational_threshold + algebraic_threshold;
+    /// 计算合并阈值（返回 uint16_t 以避免两个 uint8_t 相加溢出）
+    [[nodiscard]] uint16_t combined_threshold() const noexcept {
+        return static_cast<uint16_t>(rational_threshold) + algebraic_threshold;
     }
 };
